renamed middleware to router
