+++ api/src/app/router/error.ts
